improve state handling of handlers
